Create mod for util functions
